+++ domain/src/main/java/com/switchfully/teamteam/parkshark/domain/directors/DirectorRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DirectorRepository extends JpaRepository<Director, Long> {
+public interface DirectorRepository extends JpaRepository<Director, Integer> {
 }
