+++ domain/src/main/java/com/switchfully/teamteam/parkshark/domain/divisions/DivisionRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DivisionRepository extends JpaRepository<Division, Long> {
+public interface DivisionRepository extends JpaRepository<Division, Integer> {
 
     Division findByName(String name);
 }
